+++ src/combined/ascon-aead-masked.h
@@ -23,6 +23,8 @@
 #ifndef LWCRYPTO_ASCON_AEAD_MASKED_H
 #define LWCRYPTO_ASCON_AEAD_MASKED_H
 
+#include <stddef.h>
+
 /**
  * \file ascon-aead-masked.h
  * \brief Masked ASCON-128 encryption algorithm and related family members.
@@ -96,9 +98,9 @@ extern "C" {
  * \sa ascon128_masked_aead_decrypt()
  */
 int ascon128_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k);
@@ -126,10 +128,10 @@ int ascon128_masked_aead_encrypt
  * \sa ascon128_masked_aead_encrypt()
  */
 int ascon128_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k);
 
@@ -155,9 +157,9 @@ int ascon128_masked_aead_decrypt
  * \sa ascon128a_masked_aead_decrypt()
  */
 int ascon128a_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k);
@@ -185,10 +187,10 @@ int ascon128a_masked_aead_encrypt
  * \sa ascon128a_masked_aead_encrypt()
  */
 int ascon128a_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k);
 
@@ -214,9 +216,9 @@ int ascon128a_masked_aead_decrypt
  * \sa ascon80pq_masked_aead_decrypt()
  */
 int ascon80pq_masked_aead_encrypt
-    (unsigned char *c, unsigned long long *clen,
-     const unsigned char *m, unsigned long long mlen,
-     const unsigned char *ad, unsigned long long adlen,
+    (unsigned char *c, size_t *clen,
+     const unsigned char *m, size_t mlen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *nsec,
      const unsigned char *npub,
      const unsigned char *k);
@@ -244,10 +246,10 @@ int ascon80pq_masked_aead_encrypt
  * \sa ascon80pq_masked_aead_encrypt()
  */
 int ascon80pq_masked_aead_decrypt
-    (unsigned char *m, unsigned long long *mlen,
+    (unsigned char *m, size_t *mlen,
      unsigned char *nsec,
-     const unsigned char *c, unsigned long long clen,
-     const unsigned char *ad, unsigned long long adlen,
+     const unsigned char *c, size_t clen,
+     const unsigned char *ad, size_t adlen,
      const unsigned char *npub,
      const unsigned char *k);
